+++ tests/testthat/test-training.R
@@ -16,24 +16,6 @@ test_that("model training works with trainSLOPE", {
   }
 })
 
-test_that("plot.trainSLOPE works as expected", {
-  xy <- SLOPE:::randomProblem(1e2, 2)
-  x <- xy$x
-  y <- xy$y
-
-  fit <- trainSLOPE(x, y, solver = "admm")
-
-  expect_error(plot(fit, measure = "auc"))
-  p <- plot(fit)
-  expect_s3_class(p, "trellis")
-  expect_silent(dont_plot(p))
-
-  fit <- trainSLOPE(x, y, q = c(0.1, 0.2), number = 2, solver = "admm")
-
-  p <- plot(fit)
-  expect_s3_class(p, "trellis")
-})
-
 
 test_that("trainSLOPE works properly for binomial family", {
   set.seed(1454)